Remove redundant tests, moving only relevant ones to main test file
